+++ game_patch/hud/hud_weapons.cpp
@@ -1,5 +1,7 @@
 #include <patch_common/CallHook.h>
 #include <patch_common/FunHook.h>
+#include <patch_common/CodeInjection.h>
+#include "xlog/xlog.h"
 #include "../rf/gr/gr.h"
 #include "../rf/hud.h"
 #include "../rf/entity.h"
@@ -7,11 +9,14 @@
 #include "../rf/gr/gr_font.h"
 #include "../rf/player/player.h"
 #include "../rf/multi.h"
+#include "../graphics/gr.h"
 #include "../main/main.h"
 #include "../misc/alpine_settings.h"
+#include "../misc/misc.h"
 #include "hud_internal.h"
 
 float g_hud_ammo_scale = 1.0f;
+bool g_displaying_custom_reticle = false;
 
 CallHook<void(int, int, int, rf::gr::Mode)> hud_render_ammo_gr_bitmap_hook{
     {
@@ -55,11 +60,17 @@ CallHook<void(int, int, int, int)> render_reticle_set_color_hook{
     0x0043A4D7,
     [](int r, int g, int b, int a) {
         rf::Color clr{};
-        if (g_alpine_game_config.reticle_color_override) {
+
+        if (g_displaying_custom_reticle && !g_alpine_game_config.colorize_custom_reticles) {
+            clr = {255, 255, 255, 255}; // white
+        }
+        else if (g_alpine_game_config.reticle_color_override) {
             clr = rf::Color::from_hex(*g_alpine_game_config.reticle_color_override);
         }
         else {
-            clr = {0, 255, 0, 255};
+            clr = g_displaying_custom_reticle ?
+                rf::Color{255, 255, 255, 255} : // white
+                rf::Color{0, 255, 0, 255}; // green
         }
 
         render_reticle_set_color_hook.call_target(clr.red, clr.green, clr.blue, clr.alpha);
@@ -70,14 +81,46 @@ CallHook<void(int, int, int, int)> render_reticle_locked_set_color_hook{
     0x0043A472,
     [](int r, int g, int b, int a) {
         rf::Color clr{};
-        if (g_alpine_game_config.reticle_locked_color_override) {
+
+        if (g_displaying_custom_reticle && !g_alpine_game_config.colorize_custom_reticles) {
+            clr = {255, 255, 255, 255}; // white
+        }
+        else if (g_alpine_game_config.reticle_locked_color_override) {
             clr = rf::Color::from_hex(*g_alpine_game_config.reticle_locked_color_override);
         }
         else {
-            clr = {255, 0, 0, 255};
+            clr = g_displaying_custom_reticle ?
+                rf::Color{255, 255, 255, 255} : // white
+                rf::Color{255, 0, 0, 255}; // red
         }
 
         render_reticle_locked_set_color_hook.call_target(clr.red, clr.green, clr.blue, clr.alpha);
+    },
+};
+
+CodeInjection render_reticle_check_custom_injection{
+    0x0043A3B1,
+    [](auto& regs) {
+        int weap_slot = regs.eax;
+
+        if (weap_slot >= 0) {
+            float base_scale = g_alpine_game_config.big_hud ? 2.0f : 1.0f;
+            float scale = base_scale * g_alpine_game_config.get_reticle_scale();
+            bool big_reticle = scale > 1.0f; // reticle is using _1 variant
+            g_displaying_custom_reticle = weapon_reticle_is_customized(weap_slot, big_reticle);
+
+            // special case for rocket lock on reticle
+            if (weap_slot == rf::rocket_launcher_weapon_type) {
+                rf::Player* pp = regs.edi;
+                if (rf::player_fpgun_locked_on(pp)) {
+                    g_displaying_custom_reticle = rocket_locked_reticle_is_customized(big_reticle);
+                }
+            }
+        }
+        else {
+            g_displaying_custom_reticle = false;
+        }
+        //xlog::warn("player {}, weap {}, big? {}, custom? {}", pp->name, weap_slot, big_reticle, g_displaying_custom_reticle);
     },
 };
 
@@ -138,5 +181,5 @@ void hud_weapons_apply_patches()
     render_reticle_gr_bitmap_hook.install();
     render_reticle_set_color_hook.install();
     render_reticle_locked_set_color_hook.install();
-
+    render_reticle_check_custom_injection.install();
 }